pyproject.toml
[project]
name = "piaic_card_generator"
version = "0.1.0"
description = "piaic_card_generator using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "chainlit>=2.2.1",
    "crewai[tools]>=0.102.0,<1.0.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
plot = "piaic_card_generator.tools.main:plot"
run_crew = "piaic_card_generator.tools.main:kickoff"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/piaic_card_generator"]  

[tool.crewai]
type = "flow"
